fix: Reinclude node_modules in build-upload-bundle s3 cp
We excluded node_modules to solve an issue with a local reference
as a symlink which made s3 cp exit with a non-zero exit code.
However, we need to keep node_modules for some other stacks,
notably `UserFriendlyNotificationsStack`.
Since the `--no-follow-symlinks` is enough to avoid an error
in the original issue, let's reinclude node_modules.