ci: Add api-upgrade Workflow
This worklow is meant to be used in other repos with a schedule trigger
It will automatically upgrade the provided api (using `npm run upgrade-api:<api>`)
or the default one if none is passed (using `npm run upgrade-api`) with a patch version.
It will then open a new PR containing the upgrade.