fix(api-upgrade): Use hardcoded nautilus user instead of GITHUB_ACTOR
GITHUB_ACTOR is supposed to be set to the user who triggered the workflow.
Is the case of a scheduled workflow, it seems to be set to the user
who created it.
However, it wasn't set in a couple of repos (probably those where I
never committed) so it will be safer if we use a hardcoded user
instead.